move current 'admin' to 'settings'
