+++ frontend/config/src/routes.ts
@@ -5,6 +5,8 @@ export enum AppRoute {
   ForgotPassword = 'forgot-password',
 
   Admin = 'admin',
+
+  Settings = 'settings',
   Logout = 'logout',
 
   PageNotFound = 'page-not-found',
